update backend to working version
