+++ logic/countries.js
@@ -1,4 +1,4 @@
-import data from '../pages/api/country/_services/_countries-slim.json';
+import data from '../countryData/data/countryList.json';
 
 export default function getCountries() {
   return { data, error: null };
